Improve `ricardeos.py` script to fit with actual documentation about ricardian contracts
The documentation [1] states that ricardian contact should be in the form:

```
// Ricardian contract of the overall contract
<contract_name>_rc.md

// Ricardian contract of each action
<contract_name>.<action_name>_rc.md
```

The `scripts/ricardeos.py` was not handling this correctly anymore. Updated
it to read the actions contract files using the right pattern.

Also change the way ricardian clauses are retrieved. Instead of a single
clause file, we "parse" the actual ricardian contract file looking for each
section of the markdown file (we only accept "H3" section (`###`)) gathering
section's line along the way.

I get that this last one might be questionable. I felt however that it's
quite better for the developer to specify a single file.

* [1] - https://developers.eos.io/eosio-cpp/docs/hello-world#section-hello-world-ricardian-contract