Split monolith string_tests into many small tests
In some environments, the single huge unit test in string_tests.cpp
either failed to compile or consumed ridiculous amounts of memory and
CPU time. Because the test comprised of many independent small cases,
lumped together to avoid having to list them in the main function, it
was easy to split the single test into individual test cases, which
helped the compilation speed and memory consumption.